requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
bcrypt==4.0.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0
PyJWT==2.8.0
Werkzeug==2.3.7